fix(Formats): performance improvement by decoupling the two graphql queries
